Code cleanup in vr_environment

+++ HealthCar3/BikeApp/data/components/TransformComponent.cs
@@ -5,16 +5,16 @@
      */
     internal class TransformComponent
     {
-        public int[] position;
+        public double[] position;
         public double scale;
         public int[] rotation;
 
         /*
          * Initializes the values.
          */
-        public TransformComponent(int posX, int posY, int posZ, double scaling, int rotX, int rotY, int rotZ)
+        public TransformComponent(double posX, double posY, double posZ, double scaling, int rotX, int rotY, int rotZ)
         {
-            position = new int[] { posX, posY, posZ };
+            position = new double[] { posX, posY, posZ };
             scale = scaling;
             rotation = new int[] { rotX, rotY, rotZ };
         }
